Added tests with respect to the JIRA issue : http://tintin.cs.indiana.edu:8091/browse/FATWVTEN-109#comment-10610
Added tests for GameBase and GameProgressSaveInfo classes and cleaned up
commented code in "Resource.java". This class may be removed by tonight
as it seems unused as code from this class has been shifted to
"Resources.java". Holding up on the team's response before refactoring
and removing the unused class.